CHROMIUM: mm: make perproc-recalim THP aware
I got a report there is THP user with perproc reclaim. Thus, this patch
adds THP page handling logic. It's almost same with MADV_FREE in that
"let's do not split the THP page unless we are not sure because THP
collapsing is never cheap". To achieve it, it splits the page only if
we know there is only user for the page via page_mapcount check.

BUG=chromium:973963
TEST=build/boot/test on 4.19

Change-Id: Ia32eceb25742362033a9b365a39c06c40f3e735c
Signed-off-by: Minchan Kim <minchan@google.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Signed-off-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1894522
Reviewed-by: Yu Zhao <yuzhao@chromium.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: HeroBuxx <herobuxx@gmail.com>